requirements.txt
Django
django-heroku
djangorestframework
gunicorn
psycopg2
pyTelegramBotAPI
dj-database-url
